Final version with updated protoboffer and corresponding code
