+++ Cargo.toml
@@ -25,6 +25,10 @@ serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 reqwest = { version = "0.11.4", features = ["json"] }
 
+# TODO: necessary for building the protos
+# [build-dependencies]
+# tonic-build = "0.5"
+
 [[bin]]
 name = "server"
 path = "src/server.rs"
