--- src/build.rs
@@ -1,4 +1,0 @@
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("proto/dbscanserving.proto")?;
-    Ok(())
-}
